Moving retry logic into SqlQueueClient.
